Update log.py
Add Docker file and git workflow.

+++ tests/scenarios/src/environment/log.py
@@ -35,9 +35,6 @@ def get_keyword_from_log(node_idx, keyword, retry_time=30):
     while i < retry_time:
         i = i + 1
         log_path = f"crates/arpa-node/log/running/node{node_idx}.log"
-        # Check if file exists, if not create an empty file
-        if not os.path.exists(log_path):
-            open(log_path, 'w', encoding='UTF-8').close()
         with open(log_path, 'r', encoding='UTF-8') as process:
             log_info = get_log_info(process, keyword)
             if log_info is not None:
@@ -82,11 +79,14 @@ def clear_log(path='crates/arpa-node/log/running/'):
     """
     Clear the node log file
     """
+    os.makedirs(path, exist_ok=True)
     node_idx = 1
     while node_idx <= 10:
         open_path = path + 'node' + str(node_idx) + '.log'
-        open(open_path, 'w', encoding='UTF-8').close()
+        if os.path.exists(open_path):
+            open(open_path, 'w', encoding='UTF-8').close()
         node_idx = node_idx + 1
+    open(path + 'anvil-chain.log', 'w', encoding='UTF-8').close()
 
 def get_err_log_from_chain():
     """
